requirements.txt
peewee==3.16.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.0.3
pydantic-settings==2.0.2
aiogram==3.0.0b8